# src/main/webapp/file/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de estados",
    "description": "API para controle de estados do Brasil, suas respectivas populações e suas respectivas areas.",
    "version": "0.0.1",
    "contact": {
      "name": "Suporte",
      "email": "oswaldo.filho.pb@compasso.com.br"
    }
  },
  "externalDocs": {
    "description": "Documentação externa",
    "url": "http://github.com"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/",
      "description": "API"
    }
  ],
  "paths": {
    "/states": {
      "get": {
        "summary": "Busca todas os estados.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/States"
          }
        }
      },
      "post": {
        "summary": "Cadastra um novo estado.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "population": {
                    "type": "number"
                  },
                  "capital": {
                    "type": "string"
                  },
                  "area": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/State"
          }
        }
      }
    },
    "/states/": {
      "get": {
        "summary": "Põe filtros de região, tamanho da população e tamanho da área.",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUL",
                "SUDESTE",
                "CENTRO_OESTE",
                "NORTE",
                "NORDESTE"
              ]
            },
            "required": false
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "population,desc",
                "area,desc"
              ]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/States"
          },
          "404": {
            "description": "Estado não encontrado.",
            "content": {
              "application/json": {
                "example": "Not Found"
              }
            }
          }
        }
      }
    },
    "/states/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "number"
          },
          "required": true
        }
      ],
      "get": {
        "summary": "Busca o estado pelo ID.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/State"
          },
          "404": {
            "description": "Estado não encontrado.",
            "content": {
              "application/json": {
                "example": "Not Found"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza um estado.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "population": {
                    "type": "number"
                  },
                  "capital": {
                    "type": "string"
                  },
                  "area": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/State"
          },
          "404": {
            "description": "Estado não encontrado.",
            "content": {
              "application/json": {
                "example": "Not Found"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Exclui um estado.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/State"
          },
          "404": {
            "description": "Estado não encontrado.",
            "content": {
              "application/json": {
                "example": "Not Found"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "State": {
        "type": "object",
        "title": "State",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "population": {
            "type": "number"
          },
          "capital": {
            "type": "string"
          },
          "area": {
            "type": "number"
          }
        }
      }
    },
    "responses": {
      "State": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                },
                "region": {
                  "type": "string"
                },
                "population": {
                  "type": "number"
                },
                "capital": {
                  "type": "string"
                },
                "area": {
                  "type": "number"
                }
              }
            }
          }
        }
      },
      "States": {
        "description": "Sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/State"
              }
            }
          }
        }
      }
    }
  }
}